+++ lib/bali/form_builder/radio_fields.rb
@@ -16,7 +16,84 @@ module Bali
         field_helper(method, field, options)
       end
 
+      def radio_buttons_group(
+        method, values, options = {}, togglers_options = {}, radios_options = {}
+      )
+        @template.render Bali::FieldGroupWrapper::Component.new self, method, options do
+          radio_buttons_field(
+            method, values, options, togglers_options, radios_options
+          )
+        end
+      end
+
+      def radio_buttons_field(
+        method, values, options = {}, togglers_options = {}, radios_options = {}
+      )
+        current_value = options.delete(:current_value) ||
+                        values.find { |key, _| values[key].present? }&.first
+
+        options[:control_class] = "radio-buttons-group #{options[:control_class] || ''}"
+        options[:control_data] ||= {}
+        options[:control_data].merge!(
+          controller: 'radio-buttons-group',
+          'radio-buttons-group-current-value': current_value,
+          'radio-buttons-group-keep-selection-value': options.delete(:keep_selection)
+        )
+
+        field = safe_join([
+                            togglers(values, togglers_options),
+                            radio_buttons(method, values, radios_options)
+                          ])
+
+        field_helper(method, field, options)
+      end
+
       private
+
+      def togglers(values, options)
+        options = prepend_class_name(options, 'togglers')
+        toggler_opts = toggler_options(options)
+
+        tag.div(**options) do
+          safe_join(values.keys.map do |value|
+            toggler_opts[:disabled] = values[value].blank?
+            toggler_opts[:value] = value
+
+            tag.button(value, **toggler_opts)
+          end)
+        end
+      end
+
+      def toggler_options(options)
+        opts = options.delete(:toggler) || {}
+        opts = prepend_class_name(opts, 'toggler')
+        opts = prepend_action(opts, 'radio-buttons-group#change')
+        opts = prepend_data_attribute(opts, 'radio-buttons-group-target', 'toggler')
+        opts[:type] = 'button'
+
+        opts
+      end
+
+      def radio_buttons(method, values, options)
+        options = prepend_data_attribute(options, 'radio-buttons-group-target', 'element')
+
+        label_options = options.delete(:label) || {}
+        label_options = prepend_class_name(label_options, 'radio')
+
+        safe_join(values.map do |category, category_values|
+          options[:data]['radio-buttons-group-value'] = category
+
+          tag.div(**options) do
+            safe_join(
+              tags(
+                category_values,
+                { data: { action: 'click->radio-buttons-group#select' } },
+                method, label_options[:class]
+              )
+            )
+          end
+        end)
+      end
 
       def tags(values, html_options, method, label_class)
         field_name = [object.model_name.singular, method].join('_')