+++ spec/dummy/app/javascript/application.js
@@ -12,6 +12,7 @@ import {
   NavbarController,
   NotificationController,
   PopupController,
+  RadioButtonsGroupController,
   RateController,
   RevealController,
   SelectedController,
@@ -36,6 +37,7 @@ application.register('hovercard', HovercardController)
 application.register('navbar', NavbarController)
 application.register('notification', NotificationController)
 application.register('popup', PopupController)
+application.register('radio-buttons-group', RadioButtonsGroupController)
 application.register('rate', RateController)
 application.register('reveal', RevealController)
 application.register('selected', SelectedController)